wq and vine use resource_monitor module
