+++ taskvine/src/bindings/python3/ndcctools/taskvine/utils.py
@@ -14,21 +14,6 @@ def get_c_constant(constant):
     return getattr(cvine, constant)
 
 
-def rmsummary_snapshots(self):
-    """ Converts snapshot array from C to a python list """
-    if self.snapshots_count < 1:
-        return None
-
-    snapshots = []
-    for i in range(0, self.snapshots_count):
-        snapshot = cvine.rmsummary_get_snapshot(self, i)
-        snapshots.append(snapshot)
-    return snapshots
-
-
-cvine.rmsummary.snapshots = property(rmsummary_snapshots)
-
-
 def set_port_range(low_port, high_port):
     """ Sets the range for CCTools to look for free ports. """
     if low_port > high_port:
